requirements.txt
-i https://pypi.org/simple
apache-libcloud==3.8.0; python_version >= '3.7'
certifi==2025.4.26; python_version >= '3.6'
charset-normalizer==3.4.2; python_version >= '3.7'
idna==3.10; python_version >= '3.6'
requests==2.32.3; python_version >= '3.8'
urllib3==2.4.0; python_version >= '3.9'
